feat: scaffold rss feed setup n styls
