+++ src/components/head.astro
@@ -3,6 +3,7 @@ import { formatCanonicalURL } from "@/helpers/seo";
 import { defaultKeywords, delimiter } from "@/lib/config";
 import { getOgImageUrl } from "@/lib/og-image";
 import "@/styles/main.css";
+import "@/styles/print.css";
 // import { pwaAssetsHead } from "virtual:pwa-assets/head";
 
 type Props = Meta & { og?: OpenGraph; twitter?: Twitter };
