--- src/styles/base.css
@@ -1,226 +1,0 @@
-@import "./vars.css";
-@import "./fonts.css";
-
-:not(:defined) {
-  display: none;
-}
-
-html,
-body {
-  margin: 0;
-  padding: 0;
-  height: 100%;
-  width: 100%;
-  box-sizing: border-box;
-
-  font-family: var(--sans-serif-font-family);
-  font-size: var(--body-text);
-  font-weight: normal;
-
-  display: flex;
-  flex-direction: column;
-  justify-content: stretch;
-  align-items: stretch;
-
-  background: var(--dark-bg-color);
-  color: var(--text-color);
-
-  scrollbar-width: thin;
-  scrollbar-color: var(--theme-color) var(--scrollbar-bg);
-  overscroll-behavior-y: contain;
-}
-
-*,
-*::after,
-*::before {
-  box-sizing: inherit;
-  font-family: inherit;
-  -webkit-tap-highlight-color: transparent;
-}
-
-*:has(> svg) {
-  display: inline-flex;
-  align-items: center;
-  gap: 0.3em;
-  overflow-wrap: break-word;
-  word-break: break-word;
-}
-
-*:has(> svg) svg {
-  min-width: 1em;
-  min-height: 1em;
-}
-
-::selection {
-  background-color: var(--theme-color);
-  color: var(--bg-color);
-}
-
-::target-text {
-  background-color: var(--accent-color);
-  color: var(--bg-color);
-}
-
-:focus-visible {
-  outline: var(--theme-color) solid var(--border-width);
-  outline-offset: var(--spacing-small);
-  background-color: var(--bg-color);
-  color: var(--text-color);
-}
-
-a {
-  color: var(--link-color);
-  overflow-wrap: break-word;
-  word-break: break-word;
-  text-underline-offset: 0.15em;
-  text-decoration-skip-ink: auto;
-  text-decoration-thickness: 0.08em;
-}
-
-a:visited {
-  color: var(--link-visited-color);
-}
-a:focus-visible {
-  color: var(--text-color);
-}
-
-a:hover {
-  color: var(--link-hover-color);
-  outline-color: var(--link-hover-color);
-}
-
-a:active {
-  color: var(--link-active-color);
-  outline-color: var(--link-active-color);
-}
-
-a:has(img) {
-  display: inline-flex;
-}
-
-a:focus-visible:has(img) {
-  background-color: var(--bg-color);
-}
-
-h1 {
-  text-align: center;
-  color: var(--theme-color);
-  margin-block: 0;
-  margin-inline: auto;
-}
-
-h1 img {
-  height: 100%;
-  object-fit: contain;
-  width: clamp(var(--min-width), 70vmin, var(--medium-width));
-}
-
-h1,
-h2,
-h3,
-h4,
-h5,
-h6,
-summary {
-  font-family: var(--monospace-font-family);
-  font-weight: bold;
-  color: var(--theme-color);
-}
-
-main {
-  flex-grow: 1;
-}
-
-summary {
-  list-style: none;
-  display: flex;
-  align-items: center;
-  gap: var(--spacing-small);
-}
-
-summary::-webkit-details-marker {
-  display: none;
-}
-summary::marker {
-  display: none;
-}
-
-summary::before {
-  --icon-size: 1.2em;
-
-  content: "";
-  display: inline-block;
-  width: var(--icon-size);
-  height: var(--icon-size);
-  background-image: url("data:image/svg+xml,%3Csvg xmlns='http://www.w3.org/2000/svg' width='1em' height='1em' viewBox='0 0 24 24'%3E%3Cpath fill='%230080ff' d='M15.54 11.29L9.88 5.64a1 1 0 0 0-1.42 0a1 1 0 0 0 0 1.41l4.95 5L8.46 17a1 1 0 0 0 0 1.41a1 1 0 0 0 .71.3a1 1 0 0 0 .71-.3l5.66-5.65a1 1 0 0 0 0-1.47'/%3E%3C/svg%3E");
-  background-size: contain;
-  background-position: center;
-  background-repeat: no-repeat;
-}
-
-details[open] summary::before {
-  transform: rotate(90deg);
-}
-
-ul {
-  list-style: "◆ ";
-}
-ul ul {
-  list-style: "◇ ";
-}
-ul ul ul {
-  list-style: "✢ ";
-}
-
-ol {
-  list-style: decimal-leading-zero;
-}
-ol ol {
-  list-style: lower-latin;
-}
-ol ol ol {
-  list-style: lower-roman;
-}
-
-nav ul {
-  display: flex;
-  flex-wrap: wrap;
-  justify-content: center;
-  gap: var(--spacing-large);
-  padding: 0 var(--spacing-large);
-  margin: var(--spacing-large) auto;
-  list-style-type: "";
-  width: 100%;
-}
-
-iframe {
-  border: solid var(--border-width) var(--theme-color);
-  display: block;
-  margin-inline: auto;
-  min-width: var(--iframe-size);
-  min-height: var(--iframe-size);
-  width: clamp(var(--iframe-size), 100%, var(--max-width));
-  height: auto;
-  max-height: 100%;
-  aspect-ratio: 3 / 2;
-}
-
-.visually-hidden {
-  position: absolute;
-  width: 1px;
-  height: 1px;
-  margin: 0;
-  padding: 0;
-  overflow: hidden;
-  clip: rect(0, 0, 0, 0);
-  border: 0;
-  opacity: 0;
-  pointer-events: none;
-  z-index: -1;
-}
-
-@media (prefers-reduced-motion: no-preference) {
-  a {
-    text-decoration-style: dashed;
-  }
-}
