+++ src/styles/reset.css
@@ -20,7 +20,16 @@
     min-block-size: 100svh;
   }
 
-  :is(a, button, label, select, summary, [role="button"], [role="option"], [role="menuitem"]) {
+  :is(
+      a,
+      button,
+      label,
+      select,
+      summary,
+      [role="button"],
+      [role="option"],
+      [role="menuitem"]
+    ) {
     @apply cursor-pointer;
   }
