--- src/styles/vars.css
@@ -1,74 +1,0 @@
-:root {
-  /* Sizes */
-  --body-text: 20px;
-  --small-text: 0.75rem;
-
-  --min-width: 5rem;
-  --medium-width: 30rem;
-  --max-width: 76rem;
-
-  --iframe-size: var(--min-width);
-  --tag-size: 0.7rem;
-  --button-size: 2.5rem;
-
-  --avatar-size: calc(var(--medium-width) * 0.45);
-  --avatar-small-size: var(--min-width);
-  --avatar-size-mini: calc(var(--avatar-size) * 0.6);
-
-  /* Spacing */
-  --spacing-large: 1rem;
-  --spacing-medium: 0.5rem;
-  --spacing-small: 0.2rem;
-
-  /* Border */
-  --border-width: 0.1rem;
-
-  /* Theme colors */
-  --theme-color: #0080ff;
-  --secondary-color: lightblue;
-  --accent-color: #ff8000;
-  --complementary-color: orangered;
-
-  /* Base colors */
-  --text-color: black;
-  --bg-color: #c5cad3;
-  --dark-bg-color: #f2f7ff;
-  --border-color: dimgray;
-  --scrollbar-bg: rgba(255, 255, 255, 0.1);
-
-  /* Tag colors */
-  --tag-bg: var(--secondary-color);
-  --tag-color: var(--text-color);
-
-  /* Link colors */
-  --link-color: #002eaa;
-  --link-visited-color: #073869;
-  --link-hover-color: #c72800;
-  --link-active-color: #8f3237;
-
-  color-scheme: dark light;
-  accent-color: var(--theme-color);
-}
-
-@media (prefers-color-scheme: dark) {
-  :root {
-    /* Base colors */
-    --text-color: whitesmoke;
-    --bg-color: #1b1b2c;
-    --dark-bg-color: #00001f;
-    --border-color: #505050;
-    --scrollbar-bg: rgba(255, 255, 255, 0.1);
-
-    /* Tag colors */
-    --tag-bg: var(--secondary-color);
-    --tag-color: var(--dark-bg-color);
-
-    /* Link colors */
-    --link-color: #70b8ff;
-    --link-visited-color: lightblue;
-    --link-hover-color: var(--accent-color);
-    --link-active-color: var(--complementary-color);
-
-    color-scheme: dark;
-  }
-}
